requirements.txt
flask
gunicorn
google.generativeai
python-dotenv
pysqlite3
requests